+++ packages/akashic-cli-serve/spec/helper/SeverConfigMock.js
@@ -1,5 +1,0 @@
-const hostname = "localhost";
-const port = 12345;
-
-exports.hostname = hostname;
-exports.port = port;